Add WebGL main thread decoding + authorization
